requirements.txt
streamlit>=1.40.0
groq>=0.4.0
python-dotenv
